requirements.txt
jieba~=0.42.1
stylecloud~=0.5.2
numpy~=1.23.3
bilibili-api-python~=13.3.0